chore: refactored and improved zaak converters Kotlin code
