+++ src/main/java/net/atos/zac/app/informatieobjecten/converter/RESTZaakInformatieobjectConverter.java
@@ -5,8 +5,6 @@
 
 package net.atos.zac.app.informatieobjecten.converter;
 
-import static net.atos.zac.app.zaak.converter.RESTZaakStatusConverterKt.convertToRESTZaakStatus;
-
 import jakarta.inject.Inject;
 
 import net.atos.client.zgw.zrc.ZrcClientService;
@@ -20,6 +18,8 @@ import net.atos.zac.app.informatieobjecten.model.RESTZaakInformatieobject;
 import net.atos.zac.app.policy.converter.RESTRechtenConverter;
 import net.atos.zac.policy.PolicyService;
 import net.atos.zac.policy.output.ZaakRechten;
+
+import static net.atos.zac.app.zaak.model.RestZaakStatusKt.toRestZaakStatus;
 
 public class RESTZaakInformatieobjectConverter {
 
@@ -49,7 +49,7 @@ public class RESTZaakInformatieobjectConverter {
             if (zaak.getStatus() != null) {
                 final Status status = zrcClientService.readStatus(zaak.getStatus());
                 final StatusType statustype = ztcClientService.readStatustype(status.getStatustype());
-                restZaakInformatieobject.zaakStatus = convertToRESTZaakStatus(status, statustype);
+                restZaakInformatieobject.zaakStatus = toRestZaakStatus(status, statustype);
             }
         }
         return restZaakInformatieobject;
