+++ src/main/java/net/atos/zac/app/informatieobjecten/model/RESTZaakInformatieobject.java
@@ -8,7 +8,7 @@ package net.atos.zac.app.informatieobjecten.model;
 import java.time.LocalDate;
 
 import net.atos.zac.app.policy.model.RESTZaakRechten;
-import net.atos.zac.app.zaak.model.RESTZaakStatus;
+import net.atos.zac.app.zaak.model.RestZaakStatus;
 
 /**
  * weergave van een zaak die is gekoppeld aan een (enkelvoudig) informatieobject
@@ -17,7 +17,7 @@ public class RESTZaakInformatieobject {
 
     public String zaakIdentificatie;
 
-    public RESTZaakStatus zaakStatus;
+    public RestZaakStatus zaakStatus;
 
     public LocalDate zaakStartDatum;
