--- src/main/kotlin/net/atos/zac/app/zaak/model/RESTZaakStatus.kt
@@ -1,16 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2021 Atos, 2024 Lifely
- * SPDX-License-Identifier: EUPL-1.2+
- */
-package net.atos.zac.app.zaak.model
-
-import nl.lifely.zac.util.AllOpen
-import nl.lifely.zac.util.NoArgConstructor
-
-@AllOpen
-@NoArgConstructor
-data class RESTZaakStatus(
-    var naam: String,
-
-    var toelichting: String
-)
