+++ src/main/kotlin/net/atos/zac/app/zaak/model/RestZaak.kt
@@ -76,7 +76,7 @@ data class RestZaak(
     var registratiedatum: LocalDate?,
     var resultaat: RESTZaakResultaat?,
     var startdatum: LocalDate?,
-    var status: RESTZaakStatus?,
+    var status: RestZaakStatus?,
     var toelichting: String?,
     var uiterlijkeEinddatumAfdoening: LocalDate?,
     var uuid: UUID,
