+++ src/main/kotlin/net/atos/zac/app/zaak/model/RestZaakStatus.kt
@@ -2,16 +2,25 @@
  * SPDX-FileCopyrightText: 2021 Atos, 2024 Lifely
  * SPDX-License-Identifier: EUPL-1.2+
  */
-package net.atos.zac.app.zaak.converter
+package net.atos.zac.app.zaak.model
 
 import net.atos.client.zgw.zrc.model.Status
 import net.atos.client.zgw.ztc.model.generated.StatusType
-import net.atos.zac.app.zaak.model.RESTZaakStatus
+import nl.lifely.zac.util.AllOpen
+import nl.lifely.zac.util.NoArgConstructor
 
-fun convertToRESTZaakStatus(
+@AllOpen
+@NoArgConstructor
+data class RestZaakStatus(
+    var naam: String,
+
+    var toelichting: String
+)
+
+fun toRestZaakStatus(
     status: Status,
     statustype: StatusType
-) = RESTZaakStatus(
+) = RestZaakStatus(
     toelichting = status.statustoelichting,
     naam = statustype.omschrijving
 )
